api_project/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.0.0
dj-email-url==1.0.6
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cache-url==3.4.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
environs==9.5.0
fonttools==4.42.0
greenlet==2.0.2
gunicorn==21.2.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
Markdown==3.4.4
marshmallow==3.20.1
mypy-extensions==1.0.0
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.15.0
psycopg2==2.9.7
pycparser==2.21
pydyf==0.7.0
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
siuba==0.4.2
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.5.0
XlsxWriter==3.1.2
zipp==3.16.2
zopfli==0.2.2
